+++ apps/frontend/tailwind.config.ts
@@ -1,3 +1,4 @@
+import flowbite from 'flowbite-react/tailwind'
 import type { Config } from 'tailwindcss'
 
 const config: Config = {
@@ -6,6 +7,7 @@ const config: Config = {
     './src/pages/**/*.{js,ts,jsx,tsx,mdx}',
     './src/components/**/*.{js,ts,jsx,tsx,mdx}',
     './src/app/**/*.{js,ts,jsx,tsx,mdx}',
+    flowbite.content(),
   ],
   theme: {
     extend: {
@@ -16,6 +18,8 @@ const config: Config = {
       },
     },
   },
-  plugins: [],
+
+  // eslint-disable-next-line unicorn/prefer-module
+  plugins: [flowbite.plugin(), require('flowbite-typography')],
 }
 export default config